+++ frostmourne-monitor/src/main/java/com/autohome/frostmourne/monitor/controller/UserController.java
@@ -14,6 +14,7 @@ import com.autohome.frostmourne.spi.starter.model.UserInfo;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -47,5 +48,10 @@ public class UserController {
     @RequestMapping(value = "/teams", method = RequestMethod.GET)
     public Protocol<List<Team>> teams() {
         return frostmourneSpiApi.teams("frostmourne-monitor", null);
+    }
+
+    @RequestMapping(value = "/search", method = RequestMethod.GET)
+    public Protocol<List<UserInfo>> search(@RequestParam(value = "keyword", required = true) String keyword) {
+        return frostmourneSpiApi.search("frostmourne-monitor", keyword);
     }
 }